+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed Decals when rendering multiple camera in a single frame
 - Fixed cascade shadow count in shader
 - Fixed issue with Stacklit shader with Haze effect
+- Fixed an issue with the max sample count for the TAA
 
 ### Changed
 - Temporal Antialiasing optimization for Xbox One X